snmi/requirements.txt
tqdm
numpy
scipy
scikit-learn
opencv-python
Pillow
SimpleITK
tensorboard
matplotlib
nibabel
scikit-image